+++ test/utils/browser.ts
@@ -10,7 +10,7 @@ export interface IPage extends puppeteer.Page {
   $$text?: (selector: string, trim?: boolean) => Promise<(string|null)[]>;
   $attr?: (selector: string, attr: string) => Promise<string|null>;
   $$attr?: (selector: string, attr: string) => Promise<(string|null)[]>;
-  push?: (url: string, options?: puppeteer.DirectNavigationOptions) => Promise<puppeteer.Response>;
+  push?: (url: string, options?: puppeteer.WaitForOptions & { referer?: string }) => Promise<puppeteer.HTTPResponse>;
 }
 
 interface IBrowserOptions {
@@ -99,14 +99,25 @@ export default class Browser {
       page.$$eval(selector, (els, trim) => els.map((el) => {
         return trim ? (el.textContent || '').replace(/^\s+|\s+$/g, '') : el.textContent
       }), trim);
-    page.$attr = (selector, attr) =>
-      page.$eval(selector, (el, attr: string) => el.getAttribute(attr), attr);
-    page.$$attr = (selector, attr) =>
-      page.$$eval(
+    page.$attr = (selector, attr) => {
+      return page.$eval(
         selector,
-        (els, attr: string) => els.map(el => el.getAttribute(attr)),
+        (el: Element, ...args: unknown[]) => {
+          const [] = args;
+          return el.getAttribute(attr)
+        },
+        attr
+      )
+    };
+    page.$$attr = (selector, attr) =>{
+     return page.$$eval(
+        selector,
+        (els, ...args: unknown[]) => {
+          return els.map(el => el.getAttribute(attr))
+        },
         attr
       );
+    }
     return page;
   }
 }